refactor: extract NotificationItem to separate component file
- Create NotificationItem.tsx component following 1 file 1 component pattern
- Extract NotificationItem from notifications.tsx to improve code organization
- Update imports to use separate component file
- Follow React.FC pattern as per project guidelines
- Maintain all existing functionality and styling
